@@ -1,22 +1,16 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import Info from "./Info"
-import About from "./About"
-import Interests from './Interests';
-import Footer from './Footer';
-import reportWebVitals from './reportWebVitals';
+import React, { StrictMode } from "react";
+import "./index.css";
+import App from "./App";
+import reportWebVitals from "./reportWebVitals";
+import { Header } from "./components/Header";
+import { createRoot } from "react-dom/client";
 
-ReactDOM.render(
-  <React.StrictMode>
-    <div className='bissunessCard'>
-      <Info />
-      <About />
-      <Interests />
-      <Footer />
-    </div>
-  </React.StrictMode>,
-  document.getElementById('root')
+const root = createRoot(document.getElementById("root"));
+
+root.render(
+  <StrictMode>
+    <App></App>
+  </StrictMode>
 );
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
