@@ -1,23 +1,16 @@
-import logo from './logo.svg';
-import './App.css';
+import "./App.css";
+import { About } from "./components/About";
+import { Contacts } from "./components/Contacts";
+import { Education } from "./components/Education";
+import { Skills } from "./components/Skills";
 
 function App() {
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React 2
-        </a>
-      </header>
+    <div className="flex mt-5 flex-col items-center space-x-1.5 ">
+      <About></About>
+      <Education></Education>
+      <Skills></Skills>
+      <Contacts></Contacts>
     </div>
   );
 }
